firmware: add Arduino Micro (ATmega32U4) support behind ARDUINO_MICRO_BOARD config
Additive port of the PR #65 Arduino Micro work — keeps all existing
SAMD21 board configurations untouched. AVR-specific behavior is gated
behind #if defined(__AVR__) or #ifdef ARDUINO_MICRO_BOARD so SAMD21
users see zero behavior change.

- config.h: new ARDUINO_MICRO_BOARD option alongside the 5 existing
  SAMD21 configs. Defines NO_CAPACITIVE_TOUCH, NO_LED, omits BUTTON_PIN.
  Overrides memory slot dimensions to fit the 1024-byte EEPROM:
  3 slots x 100 transitions x 12 sec = 612 bytes.

- equal_temperament.h/.cpp: move lookup table to PROGMEM, expose via
  extern + GET_EQUAL_TEMPERAMENT_NOTE() macro. Saves ~256 bytes SRAM
  on AVR; transparent on SAMD (pgm_read_word is a no-op deref on ARM).

- settings_eeprom.cpp: conditionally include FlashStorage_SAMD.h on
  SAMD21 only; AVR uses native <EEPROM.h>. EEPROM.commit() routed
  through eeprom_commit() helper that compiles away on AVR while
  preserving SAMD Flash-cache flushing.

- polybuzzer.h/.cpp: POLYBUZZER_MAX_TONES drops to 1 on AVR for
  SRAM savings, stays at 2 on SAMD21 (preserves DisableBuzzer overlay
  feedback). Out-of-range slot requests are clamped to slot 0 on AVR
  so feedback tones still play.

- adapter.cpp/.h, keyers.cpp/.h: Tick(unsigned int) -> Tick(unsigned long).
  Critical fix for AVR where unsigned int is 16-bit and would overflow
  every ~65 seconds. DisableBuzzer() call in Tick() preserved — the
  DIT-hold-5-sec buzzer disable feature still works on all boards.

- touchbounce.h/.cpp, vail-adapter.ino: fully guard FreeTouch usage
  behind #ifndef NO_CAPACITIVE_TOUCH, consistent with TRRS Trinkey.

- memory.h: dimensions (MAX_MEMORY_SLOTS, MAX_TRANSITIONS_PER_MEMORY,
  MAX_RECORDING_DURATION_MS) made overridable by config.h so the Micro
  can shrink them without patching memory.h itself.

+++ adapter.cpp
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 #include <Keyboard.h>
 #include <MIDIUSB.h>
-#include <cstddef>
+#include <stddef.h>
 #include "keyers.h"
 #include "adapter.h"
 #include "polybuzzer.h"
@@ -824,7 +824,7 @@ break;
 }
 }
 
-void VailAdapter::Tick(unsigned int currentMillis) {
+void VailAdapter::Tick(unsigned long currentMillis) {
 // Check for dit hold during each tick
 if (this->ditIsHeld && this->buzzerEnabled) {
     unsigned long holdTime = currentMillis - this->ditHoldStartTime;
